bugfix: Batch processor did not shut down cleanly, losing metrics
When the process exited, the sending of metrics to OTLP was rudely
interrupted by `cancel`. This meant that the root span would often go
missing if you instrumented some kind of command line application with
the root span created at process creation.

This patch changes the shutdown logic to first ask the exporter to send
what it has and shut down, then forcibly kill it if that takes too long.