Set exception details when using WAI and Yesod instrumentation together (#121)
* Address warnings and HLint

* Extract isInternalError

* Set exception details when using WAI and Yesod instrumentation together

When the WAI instrumentation is used (so there is a span on the
request), this instrumentation only sets the shared attributes. Since
exceptions never reach WAI when used with Yesod (since they're caught to
error pages), this means that error traces don't include any details. By
calling `recordException` and `setStatus` here, we can address that.

Fixes #102

* Fix style

* Add comment about explicit exception marking