Improve source code attribute handling (#172)
* Improve source code attribute handling

My previous attempt at this was wrong for a few reasons.
1. I got the locations misaligned within the call stack, so we were
   giving the wrong call location.
2. It's in fact just totally unclear what to do with frozen call stacks
   in general.

This commit fixes 1 and enables a different approach to 2: get the user
to explicitly supply source code attributes if they are doing something
clever. So rather than trying to freeze the call stack to get us to do
waht you want, you can just create the source code attributes yourself
(using `callerAttributes` if you like) and add them. The main change to
enable this is just to not add source code attributes from the call
stack if there are any already present, which prevents us getting a
muddled mix of attributs if only some have been set.

* Add an example test with a span creation helper