Make BoundedMap.itemMap strict
I noticed that BoundedMap.itemMap was accumulating chains of thunks
while looking at a heap with ghc-debug.

This isn't necessarily a problem. BoundedMap will eventually get forced.
So, this cannot lead to a space leak.

However, this is unnecessarily laziness. There is a performance overhead
to creating and evaluating thunks. Just making this strict should be a
mild improvement for both allocations and runtime.